+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 ### 🧰 Bug fixes 🧰
 
 - The `featuregates` were not configured from the "--feature-gates" flag on windows service (#5060)
+- Fix Semantic Convention Schema URL definition for 1.5.0 and 1.6.1 versions (#5103)
 
 ## v0.47.0 Beta